requirements.txt
# LangGraph Workshop - Python Dependencies

# Core LangGraph
langgraph>=0.3.18
langgraph-checkpoint-sqlite>=0.1.0
langchain-core>=0.3.0

# LLM Providers (install at least one)
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0

# Observability
langsmith>=0.2.0

# Utilities
python-dotenv>=1.0.0
aiosqlite>=0.20.0
httpx>=0.27.0

# Development tools
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
ruff>=0.6.0
ipython>=8.20.0
